requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.6
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.10.0
django-cors-headers==4.7.0
django-filter==25.1
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filetype==1.2.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
kombu==5.5.4
langdetect==1.0.9
openai==1.107.3
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
